Commit #14:
Made the suns fall down instead of just appearing on the screen.
The sunflower suns, spawn correctly on the tile the sunflower is.
The only issue is that the Timer makes sunflowers spawn suns at the same
time instead of respect to when  they were placed. That will be fixed
soon.