Commit #15:
Impletmented a grid for my display class. It keeps track of what is
grass and what is dirt. Now you cant plant on dirt, and you cant plant a
seed on a cell that currently has a plant in it.

Also implemented the zombie level generation. The zombies will spawn
according to their sequence from the levels.csv.

Next steps:
Make zombies stop when they are on a plant.
Also complete everything from Nov 24.